requirements.txt
# Core Dependencies
streamlit>=1.32.0
python-dotenv>=1.0.0

# Google AI
google-generativeai>=0.4.0

# PDF Processing
PyMuPDF>=1.23.0

# Vector Search & Embeddings
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Utility Libraries
numpy>=1.24.0
pandas>=2.0.0